requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.13.0=py36h06a4308_0
aiohttp=3.7.4=py36h27cfd23_1
async-generator=1.10=pypi_0
async-timeout=3.0.1=py36h06a4308_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
blinker=1.4=py36h06a4308_0
brotlipy=0.7.0=py36h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h27cfd23_0
ca-certificates=2022.4.26=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2020.6.20=pyhd3eb1b0_3
cffi=1.14.6=py36h400218f_0
chardet=3.0.4=py36h06a4308_1003
click=8.0.1=pyhd3eb1b0_0
cloudpickle=1.6.0=py_0
coverage=5.5=py36h27cfd23_2
cryptography=3.4.7=py36hd23ed53_0
cudatoolkit=10.2.89=hfd86e86_1
cvxpy=1.1.18=pypi_0
cycler=0.10.0=py36_0
cython=0.29.24=py36h295c915_0
cytoolz=0.11.0=py36h7b6447c_0
dask-core=2021.3.0=pyhd3eb1b0_0
dataclasses=0.8=pyh4f3eec9_6
dbus=1.13.18=hb2f20db_0
decorator=5.0.9=pyhd3eb1b0_0
defusedxml=0.7.1=pypi_0
ecos=2.0.10=pypi_0
einops=0.4.1=pypi_0
entrypoints=0.4=pypi_0
expat=2.4.1=h2531618_2
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
glib=2.69.0=h5202010_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.1=py_2
grpcio=1.36.1=py36h2157cd5_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.10.0=py36hd6299e0_1
hdf5=1.10.6=hb1b8bf9_0
hdf5storage=0.1.18=pypi_0
icu=58.2=he6710b0_3
idna=2.10=pyhd3eb1b0_0
idna_ssl=1.1.0=py36h06a4308_0
imageio=2.9.0=pyhd3eb1b0_0
importlib-metadata=3.10.0=py36h06a4308_0
intel-openmp=2021.3.0=h06a4308_3350
ipykernel=5.5.6=pypi_0
ipython=7.16.3=pypi_0
ipython-genutils=0.2.0=pypi_0
jedi=0.17.2=pypi_0
jinja2=3.0.3=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=pypi_0
jupyter-client=7.1.2=pypi_0
jupyter-core=4.9.2=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
kiwisolver=1.3.1=py36h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.17.2=h4ff587b_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lpips=0.1.4=pypi_0
lz4-c=1.9.3=h2531618_0
markdown=3.3.4=py36h06a4308_0
markupsafe=2.0.1=pypi_0
matplotlib=3.3.4=py36h06a4308_0
matplotlib-base=3.3.4=py36h62a2d02_0
mistune=0.8.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
multidict=5.1.0=py36h27cfd23_2
munch=2.5.0=pyhd3eb1b0_0
nbclient=0.5.9=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.6=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=2.5=py_0
ninja=1.10.2=hff7bd54_1
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=py36_0
opencv3=3.1.0=py36_0
openh264=2.1.0=hd408876_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1o=h7f8727e_0
osqp=0.6.2.post5=pypi_0
packaging=21.3=pypi_0
pandas=1.1.5=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.7.1=pypi_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.3.1=py36h2c7a002_0
pip=21.1.3=py36h06a4308_0
prompt-toolkit=3.0.31=pypi_0
protobuf=3.17.2=py36h295c915_0
ptyprocess=0.7.0=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pygments=2.13.0=pypi_0
pyjwt=2.1.0=py36h06a4308_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py36h05f1152_2
pyrsistent=0.18.0=pypi_0
pysocks=1.7.1=py36h06a4308_0
python=3.6.13=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.8.0=py3.6_cuda10.2_cudnn7.6.5_0
pytz=2021.1=pypi_0
pywavelets=1.1.1=py36h7b6447c_2
pyyaml=5.4.1=py36h27cfd23_1
pyzmq=24.0.1=pypi_0
qdldl=0.1.5.post0=pypi_0
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
runipy=0.1.5=pypi_0
scikit-image=0.17.2=py36hdf5156a_0
scipy=1.5.2=py36h0b6359f_0
scs=3.1.0=pypi_0
setuptools=52.0.0=py36h06a4308_0
simpleitk=2.1.1.2=pypi_0
sip=4.19.8=py36hf484d3e_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
tensorboard=2.5.0=py_0
tensorboard-plugin-wit=1.6.0=py_0
tensorboardx=2.2=pyhd3eb1b0_0
testpath=0.6.0=pypi_0
tifffile=2020.10.1=py36hdd07704_2
timm=0.5.4=pypi_0
tk=8.6.10=hbc83047_0
toolz=0.11.1=pyhd3eb1b0_0
torchaudio=0.8.0=py36
torchinfo=1.5.4=pypi_0
torchvision=0.9.0=py36_cu102
tornado=6.1=py36h27cfd23_0
tqdm=4.62.3=pypi_0
traitlets=4.3.3=pypi_0
typing-extensions=3.10.0.0=hd3eb1b0_0
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.6=pyhd3eb1b0_1
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.6.3=py36h27cfd23_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
